+++ config/chainbridge-runtime-config.cere.dev.js
@@ -32,6 +32,15 @@ window.__RUNTIME_CONFIG__ = {
             isDoubleApproval: false,
           },
         ],
+        transferFallback: [
+          {
+            chainId: 1,
+            delayMs: 4 * 60 * 1000,
+            blockTimeMs: 6000,
+            pollingMinIntervalMs: 15000,
+            pollingMaxIntervalMs: 30000,
+          },
+        ],
       },
       {
         chainId: 1,
@@ -58,6 +67,15 @@ window.__RUNTIME_CONFIG__ = {
             decimals: cereTokenDecimals,
           },
         ],
+        transferFallback: [
+          {
+            chainId: 2,
+            delayMs: 3 * 60 * 1000,
+            blockTimeMs: 3000,
+            pollingMinIntervalMs: 15000,
+            pollingMaxIntervalMs: 30000,
+          },
+        ],
       },
       {
         chainId: 0,
@@ -81,6 +99,15 @@ window.__RUNTIME_CONFIG__ = {
               "0x000000000000000000000000000000c76ebe4a02bbc34786d860b355f5a5ce00",
             decimals: cereTokenDecimals,
             isDoubleApproval: false,
+          },
+        ],
+        transferFallback: [
+          {
+            chainId: 1,
+            delayMs: 6 * 60 * 1000,
+            blockTimeMs: 6000,
+            pollingMinIntervalMs: 15000,
+            pollingMaxIntervalMs: 30000,
           },
         ],
       },