+++ src/Contexts/Adaptors/EVMAdaptors/EVMDestinationAdaptorProvider.tsx
@@ -1,13 +1,16 @@
-import React from "react";
+import React, { useCallback } from "react";
 import { Bridge, BridgeFactory } from "@chainsafe/chainbridge-contracts";
 import { BigNumber } from "ethers";
 import { useEffect, useState } from "react";
-import { EvmBridgeConfig } from "../../../chainbridgeConfig";
+import {
+  EvmBridgeConfig,
+  getСhainTransferFallbackConfig,
+} from "../../../chainbridgeConfig";
 import { useNetworkManager } from "../../NetworkManagerContext";
 import { IDestinationBridgeProviderProps } from "../interfaces";
 import { DestinationBridgeContext } from "../../DestinationBridgeContext";
-
-import { getProvider } from "./helpers";
+import { getProvider, getErc20ProposalHash, VoteStatus } from "./helpers";
+import { Fallback } from "../../../Utils/Fallback";
 
 export const EVMDestinationAdaptorProvider = ({
   children,
@@ -17,10 +20,15 @@ export const EVMDestinationAdaptorProvider = ({
     destinationChainConfig,
     homeChainConfig,
     tokensDispatch,
+    transactionStatus,
     setTransactionStatus,
     setTransferTxHash,
     setDepositVotes,
     depositVotes,
+    depositRecipient,
+    depositAmount,
+    fallback,
+    setFallback,
   } = useNetworkManager();
 
   const [destinationBridge, setDestinationBridge] = useState<
@@ -38,7 +46,7 @@ export const EVMDestinationAdaptorProvider = ({
       );
       setDestinationBridge(bridge);
     }
-  }, [destinationChainConfig, destinationBridge]);
+  }, [destinationChainConfig, destinationBridge, transactionStatus]);
 
   useEffect(() => {
     if (
@@ -94,10 +102,12 @@ export const EVMDestinationAdaptorProvider = ({
             case 3:
               setTransactionStatus("Transfer Completed");
               setTransferTxHash(tx.transactionHash);
+              fallback?.stop();
               break;
             case 4:
               setTransactionStatus("Transfer Aborted");
               setTransferTxHash(tx.transactionHash);
+              fallback?.stop();
               break;
           }
         }
@@ -144,6 +154,75 @@ export const EVMDestinationAdaptorProvider = ({
     setTransferTxHash,
     tokensDispatch,
   ]);
+
+  const initFallbackMechanism = useCallback(async (): Promise<void> => {
+    const srcChainId = homeChainConfig?.chainId as number;
+    const destinationChainId = destinationChainConfig?.chainId as number;
+    const {
+      delayMs,
+      pollingMinIntervalMs,
+      pollingMaxIntervalMs,
+      blockTimeMs,
+    } = getСhainTransferFallbackConfig(srcChainId, destinationChainId);
+    const erc20ProposalHash = getErc20ProposalHash(
+      (destinationChainConfig as EvmBridgeConfig).erc20HandlerAddress,
+      depositAmount as number,
+      depositRecipient as string
+    );
+
+    const pollingIntervalMs = Math.min(
+      Math.max(pollingMinIntervalMs, 3 * blockTimeMs),
+      pollingMaxIntervalMs
+    );
+    const fallback = new Fallback(delayMs, pollingIntervalMs, async () => {
+      let res;
+      try {
+        res = await destinationBridge?.getProposal(
+          srcChainId,
+          parseInt(depositNonce as string),
+          erc20ProposalHash
+        );
+      } catch (error) {
+        console.error(error);
+      }
+
+      const status = res ? res[4] : undefined;
+      console.log("Proposal votes status", status);
+      switch (status) {
+        case VoteStatus.EXECUTED:
+          console.log("Transfer completed in fallback mechanism");
+          setTransactionStatus("Transfer Completed");
+          fallback.stop();
+          return false;
+        case VoteStatus.CANCELLED:
+          console.log("Transfer aborted in fallback mechanism");
+          setTransactionStatus("Transfer Aborted");
+          fallback.stop();
+          return false;
+        default:
+          return true;
+      }
+    });
+    fallback.start();
+    setFallback(fallback);
+  }, [
+    homeChainConfig,
+    destinationChainConfig,
+    depositRecipient,
+    depositNonce,
+    depositAmount,
+    destinationBridge,
+  ]);
+
+  useEffect(() => {
+    console.log({ transactionStatus }); // ToDo: check why get transaction status update several times on the same status
+    if (
+      transactionStatus === "In Transit" &&
+      destinationBridge &&
+      !fallback?.initialized()
+    )
+      initFallbackMechanism();
+  }, [transactionStatus, destinationBridge, fallback]);
 
   return (
     <DestinationBridgeContext.Provider